Reapplied 64-bit patch

+++ src/ttsubset/sft.c
@@ -740,7 +740,7 @@ static int GetCompoundTTOutline(TrueTypeFont *ttf, guint32 glyphID, ControlPoint
         index = GetUInt16(ptr, 2, 1);
         ptr += 4;
 
-        if (listFind(glyphlist, (void *) (int) index)) {
+        if (listFind(glyphlist, (void *) (long) index)) {
 #ifdef DEBUG
             fprintf(stderr, "Endless loop found in a compound glyph.\n");
             fprintf(stderr, "%d -> ", index);
@@ -754,7 +754,7 @@ static int GetCompoundTTOutline(TrueTypeFont *ttf, guint32 glyphID, ControlPoint
 #endif
         }
 
-        listAppend(glyphlist, (void *) (int) index);
+        listAppend(glyphlist, (void *) (long) index);
 
 #ifdef DEBUG2
         fprintf(stderr,"glyphlist: += %d\n", index);
@@ -938,7 +938,7 @@ static int GetTTGlyphOutline(TrueTypeFont *ttf, guint32 glyphID, ControlPoint **
         if (!glyphlist) {
             glyphlistFlag = 1;
             glyphlist = listNewEmpty();
-            listAppend(glyphlist, (void *) glyphID);
+            listAppend(glyphlist, (void *) (long) glyphID);
         }
         res = GetCompoundTTOutline(ttf, glyphID, pointArray, metrics, glyphlist);
         if (glyphlistFlag) {
@@ -1845,7 +1845,7 @@ int GetTTGlyphComponents(TrueTypeFont *ttf, guint32 glyphID, list glyphlist)
     if (glyphID >= ttf->nglyphs) return 0;
     ptr = glyf + ttf->goffsets[glyphID];
 
-    listAppend(glyphlist, (void *) glyphID);
+    listAppend(glyphlist, (void *) (long) glyphID);
 
     if (GetInt16(ptr, 0, 1) == -1) {
         guint16 flags, index;
